requirements.txt
mcp==0.1.0
pandas==2.1.4
openpyxl==3.1.2
openai==1.12.0
python-dotenv==1.0.0
numpy==1.26.3